requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dateutil==2.8.2
python-telegram-bot==20.3
requests==2.31.0
flask==3.0.0
flask-cors==4.0.0
openai==1.3.5
